backend/requirements.txt
# -------- Core Framework --------
django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

psycopg[binary]==3.2.9

# daphne==4.1.2
